requirements.txt
Flask>=2.3.0
opencv-contrib-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
Werkzeug>=2.3.0
